add new product admin action

+++ pages/api/products/[id].js
@@ -5,8 +5,11 @@ export default async function handler(req, res) {
 
   const { 
       method,
-      query: { id } // we'll need an ID (it's similar to params, remember?) 
+      query: { id }, // we'll need an ID (it's similar to params, remember?) 
+      cookies
     } = req;
+
+  const token = cookies.token;
 
   await dbConnect();
 
@@ -21,6 +24,10 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not Authenticated!");
+    }
+
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -33,6 +40,10 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not Authenticated!");
+    }
+    
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted!");
